added unit test.
