+++ process/block/metablock_test.go
@@ -2420,123 +2420,3 @@ func TestMetaProcessor_CreateBlockCreateHeaderProcessBlock(t *testing.T) {
 	err = mp.ProcessBlock(blkc, headerHandler, bodyHandler, func() time.Duration { return time.Second })
 	assert.Nil(t, err)
 }
-
-func TestMetaProcessor_CreateEpochStartFromMetaBlockEpochIsNotStarted(t *testing.T) {
-	t.Parallel()
-
-	arguments := createMockMetaArguments()
-	arguments.EpochStartTrigger = &mock.EpochStartTriggerStub{
-		IsEpochStartCalled: func() bool {
-			return false
-		},
-	}
-
-	mp, _ := blproc.NewMetaProcessor(arguments)
-
-	epStart, err := mp.CreateEpochStartForMetablock()
-	assert.Nil(t, err)
-
-	emptyEpochStart := block.EpochStart{}
-	assert.Equal(t, emptyEpochStart, *epStart)
-}
-
-func TestMetaProcessor_CreateEpochStartFromMetaBlockHashComputeIssueShouldErr(t *testing.T) {
-	t.Parallel()
-
-	expectedErr := errors.New("err computing hash")
-
-	arguments := createMockMetaArguments()
-	arguments.Marshalizer = &mock.MarshalizerStub{
-		// trigger an error on the Marshal method called from core's ComputeHash
-		MarshalCalled: func(obj interface{}) (i []byte, e error) {
-			return nil, expectedErr
-		},
-	}
-	arguments.EpochStartTrigger = &mock.EpochStartTriggerStub{
-		IsEpochStartCalled: func() bool {
-			return true
-		},
-	}
-
-	mp, err := blproc.NewMetaProcessor(arguments)
-	assert.Nil(t, err)
-
-	epStart, err := mp.CreateEpochStartForMetablock()
-	assert.Nil(t, epStart)
-	assert.Equal(t, expectedErr, err)
-}
-
-func TestMetaProcessor_CreateEpochStartFromMetaBlockShouldWork(t *testing.T) {
-	t.Parallel()
-
-	arguments := createMockMetaArguments()
-	arguments.EpochStartTrigger = &mock.EpochStartTriggerStub{
-		IsEpochStartCalled: func() bool {
-			return true
-		},
-	}
-
-	hash1 := []byte("hash1")
-	hash2 := []byte("hash2")
-
-	startHeaders := createGenesisBlocks(arguments.ShardCoordinator)
-	arguments.BlockTracker = mock.NewBlockTrackerMock(arguments.ShardCoordinator, startHeaders)
-
-	hdr := startHeaders[0].(*block.Header)
-	hdr.MetaBlockHashes = [][]byte{hash1, hash2}
-	hdr.Nonce = 1
-	startHeaders[0] = hdr
-
-	dPool := initMetaDataPool()
-	dPool.TransactionsCalled = func() dataRetriever.ShardedDataCacherNotifier {
-		return &mock.ShardedDataStub{}
-	}
-	dPool.HeadersCalled = func() dataRetriever.HeadersPool {
-		cs := &mock.HeadersCacherStub{}
-		cs.RegisterHandlerCalled = func(i func(header data.HeaderHandler, key []byte)) {
-		}
-		cs.GetHeaderByHashCalled = func(hash []byte) (handler data.HeaderHandler, e error) {
-			return &block.Header{
-				PrevHash:         []byte("hash1"),
-				Nonce:            1,
-				Round:            1,
-				PrevRandSeed:     []byte("roothash"),
-				MiniBlockHeaders: []block.MiniBlockHeader{{Hash: []byte("hash1"), SenderShardID: 1}},
-				MetaBlockHashes:  [][]byte{[]byte("hash1"), []byte("hash2")},
-			}, nil
-		}
-
-		cs.LenCalled = func() int {
-			return 0
-		}
-		cs.RemoveHeaderByHashCalled = func(key []byte) {}
-		cs.NoncesCalled = func(shardId uint32) []uint64 {
-			return []uint64{1, 2}
-		}
-		cs.MaxSizeCalled = func() int {
-			return 1000
-		}
-		return cs
-	}
-	arguments.DataPool = dPool
-	arguments.PendingMiniBlocks = &mock.PendingMiniBlocksHandlerStub{
-		PendingMiniBlockHeadersCalled: func(lastNotarizedHeaders []data.HeaderHandler) ([]block.ShardMiniBlockHeader, error) {
-			var hdrs []block.ShardMiniBlockHeader
-			hdrs = append(hdrs, block.ShardMiniBlockHeader{
-				Hash:            hash1,
-				ReceiverShardID: 0,
-				SenderShardID:   1,
-				TxCount:         2,
-			})
-
-			return hdrs, nil
-		},
-	}
-	mp, _ := blproc.NewMetaProcessor(arguments)
-
-	epStart, err := mp.CreateEpochStartForMetablock()
-	assert.Nil(t, err)
-	assert.NotNil(t, epStart)
-	//	assert.Equal(t, hash1, epStart.LastFinalizedHeaders[0].LastFinishedMetaBlock)
-	//	assert.Equal(t, hash2, epStart.LastFinalizedHeaders[0].FirstPendingMetaBlock)
-}
